pyproject.toml
[tool.poetry]
name = "fastgql"
version = "0.1.0"
description = ""
authors = ["altescy <altescy@fastmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.178.0"}
uvicorn = "^0.22.0"
sqlmodel = "^0.0.8"
aiosqlite = "^0.18.0"

[tool.poetry.dev-dependencies]
python-language-server = "^0.36.2"
pytest = "^7.3.1"
pysen = "^0.10.4"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.2.0"

[tool.pysen]
version = "0.10"

[tool.pysen-cli]
settings_dir = "."

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = true
mypy_preset = "strict"
line_length = 120
py_version = "py310"
[[tool.pysen.lint.mypy_targets]]
  paths = ["."]

[tool.pysen.lint.source]
  includes = ["."]
  excludes = [".venv/"]


[tool.black] # automatically generated by pysen
# pysen ignores and overwrites any modifications
line-length = 120
target-version = ["py310"]

[tool.isort] # automatically generated by pysen
# pysen ignores and overwrites any modifications
default_section = "THIRDPARTY"
ensure_newline_before_comments = true
force_grid_wrap = 0
force_single_line = false
include_trailing_comma = true
line_length = 120
multi_line_output = 3
use_parentheses = true
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
